requirements.txt
fastapi
uvicorn
gunicorn
google-generativeai
pydantic
